change serving ui to be on the same endpoint as API, changed serving API to have /api prefix in path
