+++ cmd/diagnostics/main.go
@@ -61,9 +61,9 @@ func main() {
 			StoreSession: cache,
 		})
 
-	r := http.NewServeMux()
+	//r := http.NewServeMux()
 
-	r.HandleFunc("/network", index)
+	/*r.HandleFunc("/network", index)
 	r.HandleFunc("/logs", index)
 	r.HandleFunc("/chain", index)
 	r.HandleFunc("/data", index)
@@ -71,15 +71,15 @@ func main() {
 	r.HandleFunc("/testing", index)
 	r.HandleFunc("/performance", index)
 	r.HandleFunc("/documentation", index)
-	r.HandleFunc("/admin", index)
+	r.HandleFunc("/admin", index)*/
 
-	buildHandler := http.FileServer(http.Dir("./../../web/dist"))
-	r.Handle("/", buildHandler)
-	r.Handle("/api", handlers)
+	//buildHandler := http.FileServer(http.Dir("./../../web/dist"))
+	//r.Handle("/ui/", buildHandler)
+	//r.Handle("/", handlers)
 
 	srv := &http.Server{
 		Addr:              fmt.Sprintf("%s:%d", listenAddr, listenPort),
-		Handler:           r,
+		Handler:           handlers,
 		MaxHeaderBytes:    1 << 20,
 		TLSConfig:         tlsConfig,
 		ReadHeaderTimeout: 1 * time.Minute,
@@ -90,47 +90,6 @@ func main() {
 			log.Fatal(err)
 		}
 	}()
-
-	/*r := http.NewServeMux()
-
-	r.HandleFunc("/network", index)
-	r.HandleFunc("/logs", index)
-	r.HandleFunc("/chain", index)
-	r.HandleFunc("/data", index)
-	r.HandleFunc("/debug", index)
-	r.HandleFunc("/testing", index)
-	r.HandleFunc("/performance", index)
-	r.HandleFunc("/documentation", index)
-	r.HandleFunc("/admin", index)
-
-	buildHandler := http.FileServer(http.Dir("./../../web/dist"))
-	r.Handle("/", buildHandler)
-
-	server := &http.Server{
-		Handler:      r,
-		Addr:         "127.0.0.1:8000",
-		WriteTimeout: 15 * time.Second,
-		ReadTimeout:  15 * time.Second,
-	}
-
-	go func() {
-		if err := server.ListenAndServe(); err != http.ErrServerClosed {
-			log.Fatal(err)
-		}
-	}()*/
-
-	/*http.Handle("/", web.UI)
-
-	server := &http.Server{
-		Addr:              ":8000",
-		ReadHeaderTimeout: 3 * time.Second,
-	}
-
-	go func() {
-		if err := server.ListenAndServe(); err != http.ErrServerClosed {
-			log.Fatal(err)
-		}
-	}()*/
 
 	open("https://localhost:8080")
 
@@ -145,10 +104,6 @@ func main() {
 		log.Println("Terminating eagerly.")
 		os.Exit(-int(syscall.SIGINT))
 	}
-}
-
-func index(w http.ResponseWriter, r *http.Request) {
-	http.ServeFile(w, r, "./../../web/dist/index.html")
 }
 
 // open opens the specified URL in the default browser of the user.
